+++ flags-sdk/reflag/app/favicon.ico
