+++ flags-sdk/reflag/components/product-detail-page/color-picker.tsx
