+++ flags-sdk/reflag/components/product-detail-page/product-details.tsx
