+++ flags-sdk/reflag/flags.ts
@@ -1,17 +1,17 @@
-import { bucketAdapter, type Context } from '@flags-sdk/bucket'
+import { reflagAdapter, type Context } from '@flags-sdk/reflag'
 import { flag } from 'flags/next'
 import { identify } from './lib/identify'
 
 export const showSummerBannerFlag = flag<boolean, Context>({
   key: 'summer_sale',
-  adapter: bucketAdapter.featureIsEnabled(),
+  adapter: reflagAdapter.isEnabled(),
   defaultValue: false,
   identify,
 })
 
 export const showFreeDeliveryBannerFlag = flag<boolean, Context>({
   key: 'free_delivery',
-  adapter: bucketAdapter.featureIsEnabled(),
+  adapter: reflagAdapter.isEnabled(),
   defaultValue: false,
   identify,
 })
